Almost everything is finished exept the routes for delete and edit users.
