+++ tsconfig.json
@@ -11,7 +11,7 @@
     "resolveJsonModule": true
   },
   "include": [
-    "src/**/*.ts"
+    "src/*.ts",
   ],
   "exclude": [
     "node_modules"
